requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
matplotlib>=3.2.2
opencv-python>=4.6.0
tqdm>=4.64.0
PyYAML>=5.3.1
Pillow>=7.1.2